pyproject.toml
[project]
name = "proxi"
version = "0.5.0"
description = "A general-purpose agent loop with multi-agent orchestration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "aiofiles>=24.0.0",
    "structlog>=24.0.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.106.0",
    "notion-client>=2.2.1",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "apscheduler>=3.10.0",
    "pyyaml>=6.0.0",
    "ddgs>=5.0.0",
    "requests>=2.31.0",
    "readability-lxml>=0.8.1",
    "html2text>=2024.2.26",
]

[project.scripts]
proxi = "proxi.main:main"
proxi-gateway = "proxi.gateway.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
